pyproject.toml
[project]
name = "papers"
version = "0.1.0"
description = "大体基本的なの全部入り"
authors = [
    { name = "n4okins", email = "63445867+n4okins@users.noreply.github.com" }
]
dependencies = [
    "torch==2.2.2+cu121",
    "numpy>=1.26.3",
    "jax[cuda12_pip]>=0.4.26",
    "jaxlib>=0.4.26",
    "torchvision==0.17.2+cu121",
    "arxiv<2.1.0",
    "pypdf>=4.2.0",
    "matplotlib>=3.8.4",
    "pillow>=10.2.0",
    "scipy>=1.13.0",
    "selenium==4.17.1",
    "beautifulsoup4>=4.12.3",
    "typing-extensions>=4.8.0",
    "flax[all]>=0.8.2",
    "dm-haiku>=0.0.12",
    "cupy-cuda12x>=13.1.0",
    "optuna>=3.6.1",
    "mujoco>=3.1.4",
    "gymnasium[all]>=0.29.1",
    "nvidia-dali-cuda120>=1.36.0",
    "box2d>=2.3.2",
    "pygame>=2.5.2",
    "pyglet>=2.0.15",
    "opencv-python>=4.9.0.80",
    "jraph>=0.0.6.dev0",
    "torch-geometric>=2.5.3",
    "ninja>=1.11.1.1",
    "wheel>=0.43.0",
    "torch-scatter==2.1.2+pt22cu121",
    "torch-sparse==0.6.18+pt22cu121",
    "torch-cluster==1.6.3+pt22cu121",
    "torch-spline-conv==1.2.2+pt22cu121",
    "pyg-lib==0.4.0+pt22cu121",
    "tqdm>=4.64.1",
    "pandas>=2.2.2",
    "sympy>=1.12",
    "polars>=0.20.21",
    "plotly>=5.21.0",
    "seaborn>=0.13.2",
    "django<5",
    "flask>=3.0.3",
    "fastapi>=0.110.2",
    "nuitka>=2.1.6",
    "flet<0.22.0",
    "pyqt6>=6.6.1",
    "networkx>=3.2.1",
    "pydantic>=2.7.0",
    "torchaudio==2.2.2+cu121",
    "jinja2>=3.1.2",
    "kornia>=0.7.2",
    "mpmath>=1.3.0",
    "huggingface-hub==0.17",
    "transformers[torch,flax]<=4.40",
    "openai>=1.23.1",
    "tiktoken>=0.6.0",
    "mlxtend>=0.23.1",
    "scikit-learn>=1.4.2",
    "scikit-image>=0.23.1",
    "nltk>=3.8.1",
    "spacy>=3.7.4",
    "ginza>=5.2.0",
    "mecab-python3>=1.0.9",
    "unidic-lite>=1.0.8",
    "janome>=0.5.0",
    "sudachipy>=0.6.8",
    "sudachidict-core>=20240409",
    "nagisa>=0.2.11",
    "sentencepiece>=0.2.0",
    "gensim>=4.3.2",
    "requests>=2.28.1",
    "pyzbar>=0.1.9",
    "tensorflow>=2.16.1",
    "jupyterlab>=4.1.6",
    "hydra-core>=1.3.2",
    "uvicorn>=0.29.0",
    "daphne>=4.1.2",
    "librosa>=0.10.1",
    "gunicorn>=22.0.0",
    "pyro-ppl>=1.9.0",
    "gpytorch>=1.11",
    "botorch>=0.10.0",
    "lightning>=2.2.2",
    "kaolin>=0.15.0",
    "fvcore>=0.1.5.post20221221",
    "iopath>=0.1.10",
    "channels[daphne]>=4.1.0",
    "fiftyone>=0.23.8",
    "datasets<2.19",
    "online-judge-tools>=11.5.1",
    "ray>=2.11.0",
    "line-bot-sdk<=3.11.0",
    "discord-py>=2.3.2",
    "pybullet>=3.2.6",
    "bpy>=4.0.0",
    "dm-pix>=0.4.2",
    "pyspark>=3.5.1",
    "dask[complete]>=2024.4.2",
    "dgl==2.1.0+cu121",
    "sounddevice>=0.4.6",
    "pyaudio>=0.2.14",
    "pip>=24.0",
    "numpyro[cuda]>=0.14.0",
    "gspread>=6.1.0",
    "openpyxl>=3.1.2",
    "pymc>=5.13.1",
    "blackjax>=1.1.1",
    "shapely>=2.0.4",
    "ijson>=3.2.3",
    "geopandas>=0.14.3",
    "tiledb>=0.28.0",
    "wikipedia>=1.4.0",
    "sphinx>=7.3.7",
    "pycryptodome>=3.20.0",
    "cryptography>=42.0.5",
    "pwntools>=4.12.0",
    "maturin>=1.5.1",
    "timm>=1.0.3",
    "torchprofile>=0.0.4",
    "kaggle<=1.6",
    "torchinfo>=1.8.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.4.5",
    "pytest>=8.2.1",
]

[[tool.rye.sources]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
type = "index"

[[tool.rye.sources]]
name = "pyg-lib"
url = "https://data.pyg.org/whl/torch-2.2.1+cu121.html"
type = "find-links"

[[tool.rye.sources]]
name = "jax"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
type = "find-links"

[[tool.rye.sources]]
name = "kaolin"
url = "https://nvidia-kaolin.s3.us-east-2.amazonaws.com/torch-2.1.1_cu121.html"
type = "find-links"

[[tool.rye.sources]]
name = "dgl"
url = "https://data.dgl.ai/wheels/cu121/repo.html"
type = "find-links"

[[tool.rye.sources]]
name = "numpyro"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
type = "find-links"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/papers"]
